Added new DeckPinSwitch file
This deck allows to switch a pin using a timer so at high frequency.
In this case more than 200Khz.
This deck is compatible with flow deck, since the flow deck v2 doesn't have the pin GPIOA2 used: is connected in the schematic but the resistors are missing in the real board.
Tested with position hold and it works.