Several Update. check description. TO BE TESTED
Splitted 3Anchors and 4 Anchors in the nelder mead implementation.
Changed the for cycles into multiple lines since the dimensionality of the problem is known to improve performance.
Optimized the number of iterations and the delay in the main loop of nelder mead to improve performance: from simulation 15 iterations or more seems to give not notable differences, and the vtaskdelay is now executed every 9 ms to be sure the kalman filter can update during neldear mead algorithm. 15 ms average for nelder mead  .

Added parabolic interpolation and peack calculation for the 3 point including the peack of the FFt of each  anchor signal, to improve the quality of the peack extraction.