BugFixing
changed the range where the simplex of the NM algorithm search for the solution.
change the max_iteration of the NM.
Reduced the NM position estimate outlier detection.
Add Log of the LKF position Estimation.
Set the Gain to 1000.
Increase a little the STD of the position estimation of the algorithm.